WIP: Add scaffolding for display toast on new submission
- The "Like" button doesn't work yet
- There's a memory leak if Content ever unmounts (other than when the page reloads).
  See the comment in the first useEffect in Content for details.
- Keeping everything in Content for now. If there's time and it makes sense I'll
  decompose components into separate modules.

Known issue: if a new form submission is submitted while a toast is displayed for a
previous one, the previous one will be replaced and the customer won't be able to
like it. Discuss with the team how best to handle this scenario. E.g. we could display
one toast for each submission, but then the screen could be cluttered with many toasts.