Add tests for LikedSubmissions
Using a known, constant seed with Chance in tests so that we get
repeatable results, and thus repeatable tests.

Right now, some tests are tightly coupled to the wording of customer-
facing text. Depending on the requirements, we may or may not want
this. Other options are to a) move all customer-facing text into a
separate module (probably a good idea anyway) and use text from this
module in tests, b) refactor so that we can use other testing-library
queries like getRole or getByTestId. One issue with using getByText
is the text can sometimes get broken up on rendering, which will cause
false negatives.